requirements.txt
discord
python-dotenv
requests
iso8601
schedule
